refactor(common): move KlantcontactService from Web.Server to Common
Enables reuse by both Web.Server and Poller projects.
The service only depends on IOpenKlantApiClient and ConflictException,
both already in Common.

Task: #372
